% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{R_cat}
\alias{R_cat}
\title{R_cat：用Rust實踐cat命令}
\usage{
R_cat(file, number_lines = FALSE, number_nonblank_lines = FALSE)
}
\arguments{
\item{file}{文件名}

\item{number_lines}{是否顯示行號}

\item{number_nonblank_lines}{是否顯示非空行號}
}
\value{
: NULL
}
\description{
選擇以不同的方式顯示文件的內容
}
\examples{
R_cat("colorRust-Ex.R", TRUE, FALSE)
}
