% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{R_head}
\alias{R_head}
\title{R_head：用Rust實踐head命令}
\usage{
R_head(dir, file_type = "", line = 10, byte = 0)
}
\arguments{
\item{dir}{可以是某個根目錄或指定文件}

\item{file_type}{文件類型，如txt,csv...，預設是不指定}

\item{line}{印出幾行內容，預設10}

\item{byte}{印出幾個字元，預設是0，0是全部的意思}
}
\description{
檢視檔案前幾行、前幾位元，也可以檢視多個文件
}
\examples{
R_head("colorRust-Ex.R", "", 10, 0)
}
