% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_cat}
\alias{rust_cat}
\title{rust_cast：用Rust實踐cat命令}
\usage{
rust_cat(file_name, number_lines, number_nonblank_lines)
}
\arguments{
\item{file_name}{文件名}

\item{number_lines}{是否顯示行號}

\item{number_nonblank_lines}{是否顯示非空行號}
}
\description{
選擇以不同的方式顯示文件的內容
}
\examples{
rust_cat("colorRust-Ex.R", TRUE, FALSE)
}
