% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_head}
\alias{rust_head}
\title{rust_head：用Rust實踐head命令}
\usage{
rust_head(file, type_file, n, byte)
}
\arguments{
\item{file}{可以是某個根目錄或指定文件}

\item{type_file}{文件類型，如txt,csv...，預設是不指定}

\item{n}{印出幾行內容，預設10}

\item{byte}{印出幾個字元，預設是0(0是全部的意思)}
}
\description{
檢視檔案前幾行、前幾位元，也可以檢視多個文件
}
\examples{
rust_head("colorRust-Ex.R", "", 10, 0)
}
